Settings (desktop) — splits `<brave-account-row>`
