+++ browser/resources/settings/getting_started_page/brave_account_logged_in_row.ts
@@ -1,0 +1,131 @@
+/* Copyright (c) 2026 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+import { CrLitElement, PropertyValues } from '//resources/lit/v3_0/lit.rollup.js'
+import { I18nMixinLit } from '//resources/cr_elements/i18n_mixin_lit.js'
+
+import { BraveAccountBrowserProxy } from './brave_account_browser_proxy.js'
+import { LoggedInState } from '../brave_account.mojom-webui.js'
+import { getCss } from './brave_account_logged_in_row.css.js'
+import { getHtml } from './brave_account_logged_in_row.html.js'
+
+export class BraveAccountLoggedInRowElement extends I18nMixinLit(CrLitElement) {
+  static get is() {
+    return 'brave-account-logged-in-row'
+  }
+
+  static override get styles() {
+    return getCss()
+  }
+
+  override render() {
+    return getHtml.bind(this)()
+  }
+
+  static override get properties() {
+    return {
+      browserProxy: { type: Object },
+      state: { type: Object },
+    }
+  }
+
+  accessor browserProxy!: BraveAccountBrowserProxy
+  protected accessor state!: LoggedInState
+
+  private measure?: (text: string) => number
+  private resizeObserver?: ResizeObserver
+
+  override disconnectedCallback() {
+    super.disconnectedCallback()
+
+    this.cleanUpEmailTruncation()
+  }
+
+  override updated(changedProperties: PropertyValues<this>) {
+    super.updated(changedProperties)
+
+    if ((changedProperties as Map<PropertyKey, unknown>).has('state')) {
+      this.updateEmailTruncation()
+    }
+  }
+
+  protected onLogOutButtonClicked() {
+    this.browserProxy.authentication.logOut()
+  }
+
+  private updateEmailTruncation() {
+    if (!this.resizeObserver) {
+      const emailEl =
+          this.shadowRoot?.querySelector<HTMLElement>('#email')
+      if (!emailEl) return
+
+      const canvas = document.createElement('canvas')
+      const ctx = canvas.getContext('2d')!
+      ctx.font = getComputedStyle(emailEl).font
+      this.measure = (text: string) => ctx.measureText(text).width
+
+      this.resizeObserver = new ResizeObserver(() => {
+        this.truncateEmail(emailEl)
+      })
+      this.resizeObserver.observe(emailEl)
+      this.truncateEmail(emailEl)
+    }
+  }
+
+  private truncateEmail(emailEl: HTMLElement) {
+    if (!this.measure) return
+
+    const email = this.state.email
+    if (!email) return
+
+    const availableWidth = emailEl.clientWidth
+    if (!availableWidth) return
+
+    if (this.measure(email) <= availableWidth) {
+      emailEl.textContent = email
+      return
+    }
+
+    // Use binary search (O(log n)) to find the maximum number of characters
+    // that fit within available width. Characters are split evenly between
+    // the start and end of the email with '…' in the middle.
+    const chars = Array.from(email)
+    const makeCandidate = (kept: number): string => {
+      const prefixLen = Math.ceil(kept / 2)
+      const suffixLen = Math.floor(kept / 2)
+      return chars.slice(0, prefixLen).join('') + '…' +
+             chars.slice(-suffixLen).join('')
+    }
+
+    let truncatedEmail = ''
+    for (let low = 0, high = chars.length; low < high; ) {
+      const middle = Math.ceil((low + high) / 2)
+      const candidate = makeCandidate(middle)
+      if (this.measure(candidate) <= availableWidth) {
+        truncatedEmail = candidate
+        low = middle
+      } else {
+        high = middle - 1
+      }
+    }
+
+    emailEl.textContent = truncatedEmail
+  }
+
+  private cleanUpEmailTruncation() {
+    this.measure = undefined
+    this.resizeObserver?.disconnect()
+    this.resizeObserver = undefined
+  }
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'brave-account-logged-in-row': BraveAccountLoggedInRowElement
+  }
+}
+
+customElements.define(
+  BraveAccountLoggedInRowElement.is, BraveAccountLoggedInRowElement)
