+++ browser/resources/settings/getting_started_page/brave_account_logged_out_row.css
@@ -1,0 +1,26 @@
+/* Copyright (c) 2026 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+/* #css_wrapper_metadata_start
+ * #type=style-lit
+ * #scheme=relative
+ * #import=./brave_account_row_common.css.js
+ * #include=brave-account-row-common
+ * #css_wrapper_metadata_end */
+
+localized-link {
+  --cr-link-color: var(--leo-color-text-interactive);
+}
+
+.second-row {
+  align-items: flex-start;
+  display: flex;
+  gap: var(--leo-spacing-m);
+  padding: 0 var(--leo-spacing-2xl) var(--leo-spacing-l);
+
+  .cancel-verification-button {
+    --leo-button-color: var(--leo-color-systemfeedback-error-text);
+  }
+}
