+++ browser/resources/settings/getting_started_page/brave_account_row.html.ts
@@ -5,131 +5,26 @@
 
 import { html, nothing } from '//resources/lit/v3_0/lit.rollup.js'
 
+import './brave_account_logged_in_row.js'
+import './brave_account_logged_out_row.js'
 import {
   AccountStateFieldTags,
-  LoggedOutVerificationIntent,
   whichAccountState,
 } from '../brave_account.mojom-webui.js'
-import { BraveAccountSettingsStrings } from '../brave_components_webui_strings.js'
 import { SettingsBraveAccountRowElement } from './brave_account_row.js'
 
 export function getHtml(this: SettingsBraveAccountRowElement) {
-  return html`
-    <div class="row-container">
-      ${this.state === undefined ? nothing : html`
-        ${whichAccountState(this.state) === AccountStateFieldTags.LOGGED_IN
-          ? html`
-            <div class="first-row">
-              <div class="circle">
-                <leo-icon name="social-brave-release-favicon-fullheight-color">
-                </leo-icon>
-              </div>
-              <div class="title-and-description">
-                <div class="title">
-                  ${this.i18n(
-                      BraveAccountSettingsStrings.BRAVE_ACCOUNT_TITLE)}
-                </div>
-                <div class="description">
-                  <div id="email">${this.state.loggedIn!.email}</div>
-                </div>
-              </div>
-              <leo-button kind="outline"
-                          size="small"
-                          @click=${this.onLogOutButtonClicked}>
-                <leo-icon name="outside" slot="icon-before"></leo-icon>
-                ${this.i18n(
-                    BraveAccountSettingsStrings
-                         .SETTINGS_BRAVE_ACCOUNT_LOG_OUT_BUTTON_LABEL)}
-              </leo-button>
-            </div>`
-          : !this.state.loggedOut!.verification ? html`
-            <div class="first-row">
-              <div class="circle">
-                <leo-icon name="social-brave-release-favicon-fullheight-color">
-                </leo-icon>
-              </div>
-              <div class="title-and-description">
-                <div class="title">
-                  ${this.i18n(
-                      BraveAccountSettingsStrings
-                           .SETTINGS_BRAVE_ACCOUNT_LOGGED_OUT_ROW_TITLE)}
-                </div>
-                <div class="description">
-                  ${this.i18n(
-                      BraveAccountSettingsStrings.BRAVE_ACCOUNT_DESCRIPTION)}
-                </div>
-              </div>
-              <leo-button kind="filled"
-                          size="small"
-                          @click=${this.openBraveAccountDialog}>
-                ${this.i18n(
-                    BraveAccountSettingsStrings
-                         .SETTINGS_BRAVE_ACCOUNT_GET_STARTED_BUTTON_LABEL)}
-              </leo-button>
-            </div>`
-          : html`
-            <div class="first-row">
-              <div class="circle">
-                <leo-icon name="social-brave-release-favicon-fullheight-color">
-                </leo-icon>
-              </div>
-              <div class="title-and-description">
-                <div class="title">
-                  ${this.i18n(
-                      BraveAccountSettingsStrings
-                           .SETTINGS_BRAVE_ACCOUNT_VERIFICATION_ROW_TITLE)}
-                </div>
-                <div class="description">
-                  ${this.state.loggedOut!.verification.intent
-                      === LoggedOutVerificationIntent.kResetPassword
-                      && this.state.loggedOut!.verification.verifiedEmail
-                    ? this.i18n(BraveAccountSettingsStrings
-                        .SETTINGS_BRAVE_ACCOUNT_RESET_PASSWORD_VERIFIED_ROW_DESCRIPTION)
-                    : html`
-                      <localized-link
-                          .localizedString=${`${
-                            this.i18n(this.state.loggedOut!.verification.intent
-                              === LoggedOutVerificationIntent.kResetPassword
-                              ? BraveAccountSettingsStrings
-                                  .SETTINGS_BRAVE_ACCOUNT_RESET_PASSWORD_ROW_DESCRIPTION_1
-                              : BraveAccountSettingsStrings
-                                  .SETTINGS_BRAVE_ACCOUNT_VERIFICATION_ROW_DESCRIPTION_1)} ${
-                            this.i18n(
-                              BraveAccountSettingsStrings
-                                   .SETTINGS_BRAVE_ACCOUNT_VERIFICATION_ROW_DESCRIPTION_2)} ${
-                            this.i18nAdvanced(BraveAccountSettingsStrings
-                              .SETTINGS_BRAVE_ACCOUNT_VERIFICATION_ROW_DESCRIPTION_3,
-                              {tags: ['a'], attrs: ['href']})}`}
-                          @link-clicked=${this.onResendConfirmationEmailLinkClicked}>
-                      </localized-link>`}
-                </div>
-              </div>
-            </div>
-            <div class="second-row">
-              <leo-button kind="plain"
-                          size="small"
-                          @click=${this.openBraveAccountDialog}>
-                ${this.i18n(this.state.loggedOut!.verification.intent
-                    === LoggedOutVerificationIntent.kResetPassword
-                    && this.state.loggedOut!.verification.verifiedEmail
-                    ? BraveAccountSettingsStrings
-                        .SETTINGS_BRAVE_ACCOUNT_SET_NEW_PASSWORD_BUTTON_LABEL
-                    : BraveAccountSettingsStrings
-                        .SETTINGS_BRAVE_ACCOUNT_ENTER_VERIFICATION_CODE_BUTTON_LABEL)}
-              </leo-button>
-              <leo-button kind="plain"
-                          size="small"
-                          class="cancel-registration-button"
-                          @click=${this.onCancelRegistrationButtonClicked}>
-                ${this.i18n(this.state.loggedOut!.verification.intent
-                    === LoggedOutVerificationIntent.kResetPassword
-                    ? BraveAccountSettingsStrings
-                        .SETTINGS_BRAVE_ACCOUNT_CANCEL_RESET_PASSWORD_BUTTON_LABEL
-                    : BraveAccountSettingsStrings
-                        .SETTINGS_BRAVE_ACCOUNT_CANCEL_REGISTRATION_BUTTON_LABEL)}
-              </leo-button>
-            </div>`
-        }
-      `}
-    </div>`
+  return this.state === undefined ? nothing
+    : whichAccountState(this.state) === AccountStateFieldTags.LOGGED_IN
+      ? html`
+        <brave-account-logged-in-row
+            .browserProxy=${this.browserProxy}
+            .state=${this.state.loggedIn}>
+        </brave-account-logged-in-row>`
+      : html`
+        <brave-account-logged-out-row
+            .browserProxy=${this.browserProxy}
+            .initiatingServiceName=${this.initiatingServiceName}
+            .state=${this.state.loggedOut}>
+        </brave-account-logged-out-row>`
 }
