+++ browser/resources/settings/sources.gni
@@ -139,13 +139,20 @@ brave_settings_ts_files = [
   "default_brave_shields_page/brave_adblock_browser_proxy.ts",
   "default_brave_shields_page/default_brave_shields_browser_proxy.ts",
   "getting_started_page/brave_account_browser_proxy.ts",
+  "getting_started_page/brave_account_logged_in_row.html.ts",
+  "getting_started_page/brave_account_logged_in_row.ts",
+  "getting_started_page/brave_account_logged_out_row.html.ts",
+  "getting_started_page/brave_account_logged_out_row.ts",
   "getting_started_page/brave_account_row.html.ts",
   "getting_started_page/brave_account_row.ts",
   "shortcuts_page/shortcuts_page.ts",
 ]
 
-brave_settings_local_css_files =
-    [ "getting_started_page/brave_account_row.css" ]
+brave_settings_local_css_files = [
+  "getting_started_page/brave_account_logged_in_row.css",
+  "getting_started_page/brave_account_logged_out_row.css",
+  "getting_started_page/brave_account_row_common.css",
+]
 
 if (enable_brave_vpn_wireguard) {
   brave_settings_web_component_files +=
